change multiprotocol support to use a new "protocol" option
The recently added Juniper VPN support was realized by adding a
service-type alias openconnect.nc. Thereby, the protocol was encoded in
the service-type. That is not great, because traditionally the service-type
was the D-Bus bus name of the VPN service. Nowadays, with multi-VPN
support that somewhat changed and it's used as a base to generate the
real bus name.
Another important role that the service-type plays is in the connection,
where the vpn.service-type setting associates a connection with it's plugin.

The service-type should not be hacked to encode the openconnect protocol
option. Instead, add a NM_OPENCONNECT_KEY_PROTOCOL key.

However, in nm-connection-editor's UI we don't want to show the protocol
as an option inside the VPN dialog. Instead, the two protocols
"anyconnect" and "nc" should result in two separate entires in the
"add-connection" list. This way, the user first selects to create either
an "anyconnect" or "nc" VPN connection, and then the UI doesn't let
him switch protocol anymore.

This is realized by exposing the protocol as an "add-detail". When
populating the list of VPN types, connection-editor can ask the plugin
whether it supports multiple entires. Later, when adding the connection,
the protocol field is pre-filled via the "add-detail-key".

Note that nm-openconnect is mostly agnostic to the actual protocol
value. When openconnect gains support for a new protocol, the plugin
will just support it for the most part. The only thing that is missing
is to generate an additional add-connection entry.
Note that the supported protocols are inside the .name file. So, the
user could edit the plugin configuration with a new protocol. As a
result, the new protocol would show up in the list. However, that
doesn't really fly, because the plugin cannot show a proper name and
description for the unknown protocol. Also, the .name file is really
not user-configuration but a part of the plugin's implementation.

https://bugzilla.gnome.org/show_bug.cgi?id=767197